refactor: Push トーストのペーシング状態機械を push-toast.svelte.ts へ分離 (#226)
ui/ui.svelte.ts(451行)から Push トーストのカウントダウン・ペーシング状態機械
(約190行)を新規 ui/push-toast.svelte.ts へ純移動。振る舞い不変・公開 API 不変。

- 移動: pushToastState/pushToastCountdown/PUSH_COUNTDOWN_MIN_HOLD_MS/内部キュー状態
  ＋ isCountdownDraining/drainCountdownQueue/displayPushToast 等の private helper
  ＋ showPushToast/showPushCompletionToast/showStickyPushToast/setPushToastCountdown/
    clearPushToast の public 関数
- ui.svelte.ts は push-toast.svelte.ts から同名 re-export、pull/modal は据え置き。
  18箇所の既存 import 元(App.svelte/actions 他)は無変更で動作(公開 API 不変)
- ToastState は pull と共用のため ui.svelte.ts に単一定義を残し type-only import
  (実行時 import 循環なし)
- 既存 ui.svelte.test 21件＋ペーシング消費テスト全緑=振る舞い不変の回帰検知
- architecture.md のモジュール構成を更新

ui.svelte.ts: 451 -> 275 行 / push-toast.svelte.ts: 192 行

Refs #226
